channel module OK

+++ src/main/java/com/nanda/banking/core/channel/ChannelRepository.java
@@ -2,5 +2,5 @@ package com.nanda.banking.core.channel;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChannelRepository extends JpaRepository<Channel,Long> {
+public interface ChannelRepository extends JpaRepository<Channel,Integer> {
 }
